Move the existing definitions from wandb.util to sdk.lib.paths
