+++ tests/pytest_tests/system_tests/test_artifacts/test_wandb_artifacts.py
@@ -1315,7 +1315,7 @@ def test_tracking_storage_handler():
     #
     # Empirically, this test fails with:
     #   AssertionError: assert 'some-file' == '/path/to/file.txt'
-    # But 'some-file' started out as a `name`, i.e. a util.LogicalFilePathStr,
+    # But 'some-file' started out as a `name`, i.e. a LogicalFilePathStr,
     # representing the location of the file *within the artifact*
     # rather than *on the filesystem*.
     #
