+++ wandb/sdk/interface/artifacts/artifact_manifest.py
@@ -3,7 +3,7 @@ from typing import TYPE_CHECKING, Dict, List, Mapping, Optional, Union
 
 from wandb import util
 from wandb.sdk.lib.hashutil import B64MD5, ETag, HexMD5
-from wandb.util import FilePathStr, LogicalFilePathStr, URIStr
+from wandb.sdk.lib.paths import FilePathStr, LogicalFilePathStr, URIStr
 
 if TYPE_CHECKING:
     from wandb.sdk import wandb_artifacts
